refactor: extract comment input form and logic into a new `LinkTimelineInput` component.
